Extract API error handling boundary
